services/price_predictor/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Porter Bagley <porter7678@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pydantic-settings = "^2.5.2"
hopsworks = "^3.7.0"
loguru = "^0.7.2"
pandas = "<2.2.0"
scikit-learn = "^1.5.2"
comet-ml = "^3.47.0"
xgboost = "^2.1.1"
ta-lib = "^0.4.32"
optuna = "^4.0.0"
joblib = "^1.4.2"
fastapi = "^0.115.2"
uvicorn = "^0.31.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
 